Move LatestSameMajor same major to util/versions.go
